Restructure project
Eliminates the more traditional structure, opting for a simpler package architecture
